unify filenames between python and typescript
