--- hogvm/python/vm_utils.py
@@ -1,2 +1,0 @@
-class HogVMException(Exception):
-    pass
