+++ hogvm/typescript/package.json
@@ -2,8 +2,8 @@
     "name": "@posthog/hogvm",
     "version": "1.0.7",
     "description": "PostHog HogQL Virtual Machine",
-    "types": "dist/bytecode.d.ts",
-    "main": "dist/bytecode.js",
+    "types": "dist/execute.d.ts",
+    "main": "dist/execute.js",
     "packageManager": "pnpm@8.3.1",
     "scripts": {
         "test": "jest --runInBand --forceExit",
